backend/app/requirements.txt
Django
djangorestframework
pymongo
dnspython
opencv-python
face-recognition
django-storages[boto3]
pinecone-client 
